[flang1] Fix a parallel do bug
When the stride of parallel do is negative, flang will always exit loop.
It is because check_loop_bound function always check the incr of a loop
to avoid out of loop bound. It is not necessary since __kmp_for_static_init
in openmp have do the same thing, and will change the stride to match
the actual number of iterations. The function is only needed when SCHEDULE
is enabled.