[workflows] Update GitHub workflows
Use the Ubuntu 22.04 runner and the release_16x and release_17x branches of
classic-flang-llvm-project to build and test Classic Flang. We use the
pre-built classic-flang-llvm-project for building, so we do not actually
need to install a specific version of GCC. Also simplify the logic for
finding and downloading a suitable prebuilt compiler.