[runtime] Fix -Wcast-function-type-mismatch warnings; NFC (1/4)
Reduction helpers in the Classic Flang runtime for different data types have
different function signatures, yet they are stored in tables with generic
types (e.g. void (*)())and are often compared with each other in contexts
requiring explicit casts. As of Clang 19, this type of code has resulted in
new compile-time warnings. This patch fixes the cases where the generic
void (*)() type is not used, and explicit casts are no longer sufficient to
silence the compiler.