[runtime] Fix -Wcast-function-type-mismatch warnings; NFC (2/4)
Wrapper functions exists for the gather/scatter runtime functions which have
an additional argument, but the wrappers are used interchangeably with the
actual gather/scatter runtime functions. This patch fixes the runtime function
types to declare the optional argument, so that the runtime functions and their
helpers have the same types.