[runtime] Fix -Wcast-function-type-mismatch warnings; NFC (3/4)
__abort_sig_hand() resets signal handlers before printing an error message
and calling abort(). The code used to set the sa_sigaction field to SIG_DFL,
requiring an explicit cast which now causes a warning. It is actually not
necessary to use sa_sigaction in this case; setting sa_handler to SIG_DFL
is the idiomatic way which also avoids the warning.